Improved fuzz.py.
Removed comparison methods, since the sort order of differing types is
undefined in Python 2.  Consequently, comparing an integer with a blist
may have a different result than comparing an integer with a list, for
example.

Added the ability to set the random number generator seed via a
command line argument.