remove __setitem__ from sortedlist docs
It's not available in the class in v1.3.6, which jives with my impression that it's not meant to be. 

On the other hand, I can imagine you might have written an optimization of 
```
__setitem__(L,i,x):
    Throw exception if i is not the insertion position of x. 
    Otherwise: 
    del L[i]
    L.add(x)
```
And in fact I'd have used such a method, in a situation where the insertion position was already deduced for some reason other than inserting `x`.